Move resource creation to a separate module
