+++ pkg/controllers/cassandra/service/resource.go
@@ -1,0 +1,34 @@
+package service
+
+import (
+	v1alpha1 "github.com/jetstack-experimental/navigator/pkg/apis/navigator/v1alpha1"
+	"github.com/jetstack-experimental/navigator/pkg/controllers/cassandra/util"
+	apiv1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/intstr"
+)
+
+func ServiceForCluster(
+	cluster *v1alpha1.CassandraCluster,
+) *apiv1.Service {
+	svc := apiv1.Service{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:            util.ResourceBaseName(cluster),
+			Namespace:       cluster.Namespace,
+			Labels:          util.ClusterLabels(cluster),
+			Annotations:     make(map[string]string),
+			OwnerReferences: []metav1.OwnerReference{util.NewControllerRef(cluster)},
+		},
+		Spec: apiv1.ServiceSpec{
+			Type: apiv1.ServiceTypeClusterIP,
+			Ports: []apiv1.ServicePort{
+				{
+					Name:       "transport",
+					Port:       int32(9300),
+					TargetPort: intstr.FromInt(9300),
+				},
+			},
+		},
+	}
+	return &svc
+}
